main update

+++ libasm.h
@@ -1,7 +1,7 @@
 #ifndef LIBASM_H
 #define LIBASM_H
 
-int		ft_strlen(char const *str);
+size_t		ft_strlen(char const *str);
 char    *ft_strcpy(char *dest, const char *src);
 int     ft_strcmp(const char *s1, const char *s2);
 ssize_t ft_write(int fd, const void *buf, size_t count);
